components: 8bit-ify what can be
Faster and saves a couple of bytes PROGMEM

+++ microflo/components/SerialIn.hpp
@@ -2,7 +2,7 @@ class SerialIn : public SingleOutputComponent {
 public:
     virtual void process(Packet in, MicroFlo::PortId port) {
         // FIXME: make device and baudrate configurable
-        const int serialDevice = -1;
+        const int8_t serialDevice = -1;
 
         if (in.isSetup()) {
             // FIXME: do based on input data instead of hardcode
