+++ microflo/components/SerialOut.hpp
@@ -2,7 +2,7 @@ class SerialOut : public SingleOutputComponent {
 public:
     virtual void process(Packet in, MicroFlo::PortId port) {
         // FIXME: make device and baudrate configurable
-        const int serialDevice = -1;
+        const int8_t serialDevice = -1;
 
         if (in.isSetup()) {
             io->SerialBegin(serialDevice, 9600);
